+++ src/tarski/search/applicability.py
@@ -5,6 +5,7 @@ from ..evaluators.simple import evaluate
 
 
 def is_applicable(model, operator):
+    """ Check whether a given (ground) operator is applicable in the given model (state). """
     return evaluate(operator.precondition, model)
 
 
@@ -31,7 +32,7 @@ def apply_effect(model, effect):
         model.remove(effect.atom.predicate, *effect.atom.subterms)
 
     elif isinstance(effect, FunctionalEffect):
-        model.setx(effect.lhs, evaluate(effect.rhs, model))
+        model.set(effect.lhs, evaluate(effect.rhs, model))
 
     else:
         raise RuntimeError(f'Don\'t know how to apply effect "{effect}"')