+++ tests/grounding/test_lp_grounding.py
@@ -10,7 +10,7 @@ from tests.common.benchmarks import get_lenient_benchmarks
 
 from tests.common.gripper import create_sample_problem
 from tests.common.simple import create_simple_problem
-from ..io.common import reader, collect_strips_benchmarks
+from ..io.common import reader, collect_strips_benchmarks, parse_benchmark_instance
 
 if shutil.which("gringo") is None:
     pytest.skip('Install the Clingo ASP solver and put the "gringo" binary on your PATH in order to test ASP-based '
@@ -164,7 +164,7 @@ def test_ground_actions_on_negated_preconditions2():
 
 def test_action_grounding_on_orgsynth():
     # Regression test for issue #82
-    problem = generate_strips_problem("organic-synthesis-opt18-strips:p01.pddl")
+    problem = parse_benchmark_instance("organic-synthesis-opt18-strips:p01.pddl")
 
     lp, tr = create_reachability_lp(problem, ground_actions=False)
     ruleset = set(lp.rules)
@@ -184,15 +184,10 @@ def test_ignore_unused_types():
     ruleset2 = set(lp.rules)
     assert ruleset == ruleset2
 
-    problem = generate_strips_problem("organic-synthesis-opt18-strips:p01.pddl")
+    problem = parse_benchmark_instance("organic-synthesis-opt18-strips:p01.pddl")
     lp, tr = create_reachability_lp(problem, ground_actions=True)
     ruleset = set(lp.rules)
     assert not any(s.startswith("action_oxidationofborane") for s in ruleset)
 
     actions = compute_action_groundings(problem)
     assert actions['oxidationofborane'] == set()
-
-
-def generate_strips_problem(tag):
-    instance_file, domain_file = collect_strips_benchmarks([tag])[0]
-    return reader().read_problem(domain_file, instance_file)