+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Added
+  - Added some basic forward search capabilities.
+
 ### Removed
 ### Deprecated
+    - Model.set() is now deprecated
+
 ### Fixed
  - Fixed a bug in `check_hypergraph_acyclicity` reported by @abcorrea.